requirements.txt
gradio==5.11.0
langchain==0.3.14
langchain_core==0.3.29
langchain_openai==0.2.14
langgraph==0.2.61
matplotlib==3.10.0
pandas==2.2.3
pydantic==2.10.5
python-dotenv==1.0.1
scanpy==1.10.4
